refactor(xapiri): extract dashboard_cost_helpers.go — cost presets + kickRefreshCmd + period helpers
costWindowPreset, costWindows, costDefaultPeriodIdx, costMonthSecs, kickRefreshCmd,
sortedCostRows, activeCostWindow, costForPeriod, and formatCost moved per ADR 0014.
